requirements_personal.txt
alabaster==0.7.12
arabic-reshaper==2.1.3
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.10.0
bootstrap-py==1.0.2
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.9
django-bootstrap-modal-forms==2.2.0
django-bootstrap-v5==1.0.7
django-jquery==3.1.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-weekday-field==1.1.0
docutils==0.17.1
fanstatic==1.1
fonttools==4.28.5
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
gobject==0.1.0
html5lib==1.1
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
js.bootstrap==3.3.4
js.jquery==3.3.1
MarkupSafe==2.0.1
packaging==21.3
pguard==0.4.0
Pillow==9.0.0
pycparser==2.21
pydyf==0.1.2
Pygments==2.10.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.5
requests==2.26.0
shutilwhich==1.1.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tinycss2==1.1.1
typing_extensions==4.0.0
urllib3==1.26.7
weasyprint==54.0
webencodings==0.5.1
WebOb==1.8.7
xhtml2pdf==0.2.5
zopfli==0.1.9
